requirements.txt
Chameleon==2.9.2
Jinja2==2.6
Mako==0.7.2
MarkupSafe==0.15
MySQL-python==1.2.3
PasteDeploy==1.5.0
Pygments==1.5
SQLAlchemy==0.7.8
WebOb==1.2.2
psycopg2==2.4.5
pyramid==1.3.3
pyramid-debugtoolbar==1.0.2
pyramid-tm==0.5
repoze.lru==0.6
simplejson==2.6.1
transaction==1.3.0
translationstring==1.1
venusian==1.0a6
virtualenv==1.7.2
waitress==0.8.1
wsgiref==0.1.2
zope.deprecation==4.0.0
zope.interface==4.0.1
zope.sqlalchemy==0.7.1
